Moved MoviesAdapter
